// web/lib/api/generated/model/shopContact.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { ShopContactChannelEnum } from './shopContactChannelEnum';
import type { ShopContactOutcomeEnum } from './shopContactOutcomeEnum';

export interface ShopContact {
  readonly id: string;
  readonly created_at: string;
  readonly channel: ShopContactChannelEnum;
  readonly outcome: ShopContactOutcomeEnum;
  readonly note: string;
  readonly by_name: string;
}